requirements.txt
SQLAlchemy==2.0.42
python-dotenv==1.1.1
pydantic==2.11.7
pytesseract==0.3.13
pdf2image==1.17.0
PyPDF2==3.0.1
pandas==2.3.1
RapidFuzz==3.13.0
pdfplumber==0.11.6
torch==2.7.1
sentence-transformers==5.0.0
openpyxl==3.1.5
transformers
psycopg2-binary==2.9.10
asyncpg==0.30.0

ollama==0.5.4
langchain==0.3.27
langchain-core==0.3.72
pydantic-settings==2.11.0
pgvector==0.4.1
inflect==7.5.0
numpy==2.3.3